refactor(move): Use hook instead hoc into Topbar
