+++ src/drive/web/modules/move/Topbar.jsx
@@ -1,43 +1,28 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import get from 'lodash/get'
-import uniqBy from 'lodash/uniqBy'
+
+import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
+import useBreakpoints from 'cozy-ui/transpiled/react/providers/Breakpoints'
+import { useQuery } from 'cozy-client'
+
 import Topbar from 'drive/web/modules/layout/Topbar'
-import { withBreakpoints } from 'cozy-ui/transpiled/react'
 import Breadcrumb from 'drive/web/modules/navigation/Breadcrumb/Breadcrumb'
 import BackButton from 'components/Button/BackButton'
-import { ROOT_DIR_ID } from 'drive/constants/config'
+import { getBreadcrumbPath } from 'drive/web/modules/move/helpers'
+import { buildOnlyFolderQuery } from 'drive/web/modules/queries'
 
-const getBreadcrumbPath = (t, displayedFolder) =>
-  uniqBy(
-    [
-      {
-        _id: ROOT_DIR_ID
-      },
-      {
-        _id: get(displayedFolder, 'dir_id')
-      },
-      {
-        _id: displayedFolder._id,
-        name: displayedFolder.name
-      }
-    ],
-    '_id'
+const MoveTopbar = ({ navigateTo, folderId }) => {
+  const { t } = useI18n()
+  const { isMobile } = useBreakpoints()
+  const folderQuery = buildOnlyFolderQuery(folderId)
+  const { fetchStatus, data } = useQuery(
+    folderQuery.definition,
+    folderQuery.options
   )
-    .filter(({ _id }) => Boolean(_id))
-    .map(breadcrumb => ({
-      _id: breadcrumb._id,
-      name:
-        breadcrumb.name ||
-        (breadcrumb._id === ROOT_DIR_ID ? t('breadcrumb.title_drive') : '…')
-    }))
 
-const MoveTopbar = (
-  { navigateTo, currentDir, fetchStatus, breakpoints: { isMobile } },
-  { t }
-) => {
-  const path = fetchStatus === 'loaded' ? getBreadcrumbPath(t, currentDir) : []
+  const path = fetchStatus === 'loaded' ? getBreadcrumbPath(t, data) : []
   const showPreviousButton = path.length > 1 && isMobile
+
   return (
     <Topbar hideOnMobile={false}>
       {showPreviousButton && (
@@ -55,13 +40,7 @@ const MoveTopbar = (
 
 MoveTopbar.propTypes = {
   navigateTo: PropTypes.func.isRequired,
-  currentDir: PropTypes.object,
-  fetchStatus: PropTypes.string.isRequired,
-  breakpoints: PropTypes.shape({
-    isMobile: PropTypes.bool
-  }).isRequired
+  folderId: PropTypes.string.isRequired
 }
-MoveTopbar.contextTypes = {
-  t: PropTypes.func.isRequired
-}
-export default withBreakpoints()(MoveTopbar)
+
+export default MoveTopbar
