+++ src/drive/web/modules/move/helpers.js
@@ -1,13 +1,15 @@
+import get from 'lodash/get'
+import uniqBy from 'lodash/uniqBy'
+
 import Alerter from 'cozy-ui/transpiled/react/deprecated/Alerter'
-// eslint-disable-next-line no-unused-vars
-const { IOCozyFile } = require('cozy-client/dist/types')
 
 import { CozyFile } from 'models'
 import logger from 'lib/logger'
+import { ROOT_DIR_ID } from 'drive/constants/config'
 
 /**
  * Returns whether one of the targeted folders is part of the current folder
- * @param {IOCozyFile[]} targets - List of folders
+ * @param {import('cozy-client/types').IOCozyFile[]} targets - List of folders
  * @param {string} currentDirId - The id of the current folder
  * @returns {boolean} - Whether one of the targeted folders is part of the current folder
  */
@@ -21,8 +23,8 @@ export const areTargetsInCurrentDir = (targets, currentDirId) => {
 /**
  * Cancel file movement function
  * @param {object} client - The CozyClient instance
- * @param {IOCozyFile[]} entries - List of files moved
- * @param {IOCozyFile[]} trashedFiles - List of ids for files moved to the trash
+ * @param {import('cozy-client/types').IOCozyFile[]} entries - List of files moved
+ * @param {import('cozy-client/types').IOCozyFile[]} trashedFiles - List of ids for files moved to the trash
  * @param {Function} registerCancelable - Function to register the promise
  * @param {Functione} refreshSharing - Function refresh sharing state
  */
@@ -72,7 +74,7 @@ export const cancelMove = async ({
 
 /**
  * Gets a name for the entry if there is only one, or a sentence with the number of elements if there are several
- * @param {IOCozyFile[]} entries - List of files moved
+ * @param {import('cozy-client/types').IOCozyFile[]} entries - List of files moved
  * @param {Function} t - Translation function
  * @returns {string} - Name for entries
  */
@@ -92,7 +94,7 @@ export const getEntriesName = (entries, t) => {
 
 /**
  * Returns whether one of the entries that is shared not only by link
- * @param {IOCozyFile[]} entries - List of files moved
+ * @param {import('cozy-client/types').IOCozyFile[]} entries - List of files moved
  * @param {Object<string, SharedDoc>} byDocId - Object with shared files by id from cozy-sharing
  * @returns {boolean} - Whether one of the entries that is shared not only by link
  */
@@ -110,3 +112,27 @@ export const hasOneOfEntriesShared = (entries, byDocId) => {
   })
   return sharedEntries.length > 0
 }
+
+export const getBreadcrumbPath = (t, displayedFolder) =>
+  uniqBy(
+    [
+      {
+        _id: ROOT_DIR_ID
+      },
+      {
+        _id: get(displayedFolder, 'dir_id')
+      },
+      {
+        _id: displayedFolder._id,
+        name: displayedFolder.name
+      }
+    ],
+    '_id'
+  )
+    .filter(({ _id }) => Boolean(_id))
+    .map(breadcrumb => ({
+      _id: breadcrumb._id,
+      name:
+        breadcrumb.name ||
+        (breadcrumb._id === ROOT_DIR_ID ? t('breadcrumb.title_drive') : '…')
+    }))
